requirements.txt
# Core Django - Latest stable versions
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2

# Database
psycopg2-binary==2.9.9

# Authentication & Security
python-decouple==3.8
django-allauth==0.61.1
django-axes==6.4.0
django-ratelimit==4.1.0

# File handling & Images
Pillow==10.3.0
django-cleanup==8.1.0

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

# Rich Text Editor
django-ckeditor==6.7.1

# Payments
paystackapi==2.1.3

# QR Codes & PDF generation
qrcode==7.4.2
reportlab==4.2.0
pdfkit==1.0.0

# Data export
openpyxl==3.1.2
django-import-export==4.0.2

# Task Queue & Caching
celery==5.3.6
redis==5.0.4
django-redis==5.4.0

# Email & Notifications
django-anymail==10.3

# API Documentation
drf-yasg==1.21.7
drf-spectacular==0.27.2

# Performance monitoring
django-debug-toolbar==4.3.0
django-silk==5.1.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0
uvicorn==0.29.0

# Environment variables
python-dotenv==1.0.1
urllib3==2.0.7
chardet==5.2.0
charset-normalizer==3.3.2
